+++ R/unpack-tspec.R
@@ -1,23 +1,74 @@
+#' Unpack a tibblify specification
+#'
+#' @param spec A tibblify specification.
+#' @param ... These dots are for future extensions and must be empty.
+#' @param fields A string of the fields to unpack.
+#' @param recurse Should unpack recursively?
+#' @param names_sep If `NULL`, the default, the inner names of fields are used.
+#'   If a string, the outer and inner names are pasted together, separated by
+#'   `names_sep`.
+#' @param names_repair Used to check that output data frame has valid names.
+#'   Must be one of the following options:
+#'
+#' * `"unique"` or `"unique_quiet"`: (the default) make sure names are unique and not empty,
+#'
+#' * `"universal" or `"unique_quiet"`: make the names unique and syntactic
+#'
+#' * `"check_unique"`: no name repair, but check they are unique,
+#'
+#' * a function: apply custom name repair.
+#'
+#' See [vctrs::vec_as_names()] for more information.
+#'
+#' @return A tibblify spec.
+#' @export
+#'
+#' @examples
+#' spec <- tspec_df(
+#'   tib_lgl("a"),
+#'   tib_row("x", tib_int("b"), tib_chr("c")),
+#'   tib_row("y", tib_row("z", tib_chr("d")))
+#' )
+#'
+#' unpack_tspec(spec)
+#' # only unpack `x`
+#' unpack_tspec(spec, fields = "x")
+#' # do not unpack the fields in `y`
+#' unpack_tspec(spec, recurse = FALSE)
 unpack_tspec <- function(spec,
                          ...,
                          fields = NULL,
                          recurse = TRUE,
                          names_sep = NULL,
-                         names_repair = "check_unique") {
-  fields_to_unpack <- check_unpack_cols(fields, spec)
+                         names_repair = c("unique", "universal", "check_unique", "unique_quiet", "universal_quiet")) {
+  rlang::check_dots_empty()
+  check_character(fields, allow_null = TRUE)
+  check_bool(recurse)
+  check_string(names_sep, allow_null = TRUE)
 
-  browser()
-  # spec$fields[fields_to_unpack] <-
-  purrr::map(
-    spec$fields[fields_to_unpack],
-    function(field) {
-      unpack_fields(
-        ,
-        names_sep,
-        recurse = recurse
+  fields_to_unpack <- check_unpack_cols(fields, spec)
+  names_repair <- arg_match(names_repair)
+  error_call <- current_call()
+
+  spec$fields <- purrr::imap(
+    spec$fields,
+    function(field, name) {
+      if (!name %in% fields_to_unpack) {
+        return(set_names(list(field), name))
+      }
+
+      unpack_field(
+        field,
+        recurse = recurse,
+        name = name,
+        names_sep = names_sep,
+        names_repair = names_repair,
+        error_call = error_call
       )
     }
   )
+
+  spec$fields <- unchop_fields(spec$fields, names_repair, error_call)
   spec
 }
 
@@ -36,40 +87,45 @@ check_unpack_cols <- function(fields, spec, error_call = caller_env()) {
   fields
 }
 
-unpack_fields <- function(fields_spec, names_sep, recurse) {
-  fields_out <- list()
-  names <- character()
-
-  for (i in seq_along(fields_spec)) {
-    field_i <- fields_spec[[i]]
-    name_i <- names(fields_spec)[[i]]
-
-    if (recurse && field_i$type %in% c("row", "df")) {
-      field_i$fields <- unpack_fields(
-        field_i$fields,
-        names_sep = names_sep,
-        recurse = recurse
-      )
-    }
-
-    if (field_i$type == "row") {
-      row_fields <- purrr::map(field_i$fields, function(row_field) {
-        row_field$key <- c(field_i$key, row_field$key)
-        row_field
-      })
-      fields_out <- c(fields_out, unname(row_fields))
-
-      if (is.null(names_sep)) {
-        names <- c(names, names(row_fields))
-      } else {
-        nms <- paste0(name_i, names_sep, names(row_fields))
-        names <- c(names, nms)
+unpack_field <- function(field_spec, recurse, name, names_sep, names_repair, error_call) {
+  if (recurse && field_spec$type %in% c("row", "df")) {
+    field_spec$fields <- purrr::imap(
+      field_spec$fields,
+      function(field, name) {
+        unpack_field(
+          field,
+          recurse = recurse,
+          name = name,
+          names_sep = names_sep,
+          names_repair = names_repair,
+          error_call = error_call
+        )
       }
-    } else {
-      fields_out <- c(fields_out, list(field_i))
-      names <- c(names, name_i)
-    }
+    )
+
+    field_spec$fields <- unchop_fields(field_spec$fields, names_repair, error_call)
   }
 
-  set_names(fields_out, names)
+  if (field_spec$type != "row") {
+    return(set_names(list(field_spec), name))
+  }
+
+  row_fields <- purrr::map(field_spec$fields, function(row_field) {
+    row_field$key <- c(field_spec$key, row_field$key)
+    row_field
+  })
+
+  if (is.null(names_sep)) {
+    row_fields
+  } else {
+    names <- paste0(name, names_sep, names(row_fields))
+    set_names(row_fields, names)
+  }
+}
+
+unchop_fields <- function(fields, names_repair, error_call) {
+  fields <- vctrs::list_unchop(fields, name_spec = "{inner}")
+  nms <- names(fields)
+  nms <- vctrs::vec_as_names(nms, repair = names_repair, call = error_call)
+  set_names(fields, nms)
 }