+++ tests/testthat/test-unpack-tspec.R
@@ -26,23 +26,6 @@ test_that("can unpack spec", {
       )
     )
   )
-
-  spec <- tspec_df(
-    tib_lgl("a"),
-    tib_recursive("x", tib_int("b"), tib_chr("c"), .children = "children")
-  )
-
-  expect_equal(
-    unpack_tspec(spec),
-    tspec_df(
-      tib_lgl("a"),
-      tib_df(
-        "x",
-        b = tib_int(c("y", "b")),
-        c = tib_chr(c("y", "c"))
-      )
-    )
-  )
 })
 
 test_that("can use names_sep", {
@@ -89,7 +72,7 @@ test_that("can recursively unpack", {
   )
 })
 
-test_that("do not unpack if not in `fields`", {
+test_that("only unpack field in `fields`", {
   spec <- tspec_df(
     tib_lgl("a"),
     tib_row("x", tib_int("b")),
@@ -114,6 +97,11 @@ test_that("do not unpack if not in `fields`", {
     )
   )
 
+  expect_snapshot({
+    (expect_error(unpack_tspec(spec, fields = "not-there")))
+    (expect_error(unpack_tspec(spec, fields = c("not-there", "also-not-there"))))
+  })
+
   # works together with `recurse`
   spec <- tspec_df(
     tib_lgl("a"),
@@ -131,3 +119,26 @@ test_that("do not unpack if not in `fields`", {
   )
 })
 
+test_that("names are repaired", {
+  spec <- tspec_df(
+    tib_lgl("a"),
+    tib_row("x", tib_int("a")),
+    tib_row("y", tib_int("b"), tib_row("z", tib_chr("b"))),
+  )
+
+  expect_snapshot({
+    # `minimal` isn't supported
+    (expect_error(unpack_tspec(spec, names_repair = "minimal")))
+    (expect_error(unpack_tspec(spec, names_repair = "check_unique")))
+  })
+
+  expect_equal(
+    unpack_tspec(spec, names_repair = "unique_quiet"),
+    tspec_df(
+      a...1 = tib_lgl("a"),
+      a...2 = tib_int(c("x", "a")),
+      b...3 = tib_int(c("y", "b")),
+      b...4 = tib_chr(c("y", "z", "b")),
+    )
+  )
+})